requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
boto3==1.37.2
botocore==1.37.2
branca==0.8.1
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
filelock==3.17.0
folium==0.19.4
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.24.1
networkx==3.2.1
numpy==1.26.4
openai==1.60.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pinecone==6.0.1
pinecone-plugin-interface==0.0.7
protobuf==5.29.3
pyarrow==19.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.11.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.4.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.41.1
streamlit_folium==0.24.0
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
torch==2.2.2
tornado==6.4.2
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
xyzservices==2025.1.0
